added readme file generation
